backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.9
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.11
supabase
pytest
httpx
pytest-asyncio
requests
google-genai